[LKE-362] add a possibility to deploy helm charts to LKE clusters
This patch implements a new function `newChartDeployerLKE` which creates
an instance which is able to push charts to LKE clusters. The kube
config file with admin privileges is saved to a temporary file which is
removed automatically when not needed anymore.

The clusteractuator.go uses this interface to apply charts from the
`charts/lkecluster/kubelet-resources/` directory describing resources
for kubelets running on child clusters.