[LKE-300] Deploy CSI for LKE clusters
We are deploying CSI in the LKE cluster itself. Note that in the future
the `csi-linode-controller` StatefulSet can be moved to the CPC cluster
in order to increase the overall reliability, for this reason the chart
is kept in the `"csi/lke"` subdirectory (with the thoughts on creating
a `"csi/cpc"` subdirectory in the same directory).

One implementation detail: the CSI requires a `linode` secret containing
an API token for both — controller and driver. We are copying this secret
from the CPC cluster.