[LKE-292] create nodes with appropriate WG configuration
Not a very small patch, but it does one simple thing: on a node
deployment it passes an API server wireguard public key to the
stackscript. The key is created and stored in an instance of
`linodeclusterconfig` CRD if not yet exists.

The stackscript then uses this key to set up wireguard configuration on
the node. On a successful setup it annotates the node with the following
records:

   * lke.linode.com/wgpub: wireguard public key of the node
   * lke.linode.com/wgip: wireguard IP address

The WG IP address added for convenience and for reference, as it is
automatically derived from the PodCidr:

   `10.2.X.0/24` -> `172.31.X.1`

The `172.31.255.1` IP address is reserved for the API server.