Rename charts/lkecluster to charts/bleeding
The new name convention will be the charts/<version> for a particular
version of an LKE cluster deployment scheme.  The "bleeding" version is
the deployment version: all the charts for such clusters will be always
updated on controller restarts and on any changes applied to cluster
objects.

The version should be specified as the "lke.linode.com/caplke-version"
annotation for a cluster object.