Reimplement reconciliation: a huge patch
0. All charts are now version-based. The cluster won't be deployed until
it has a lke.linode.com/caplke-version annotation. Currently, only the
"bleeding" version is supported.

1. Add a per-chart-set configuration file charts/<version>/config.json.
This file lists charts to-be-deployed to the kube-system-<name>
namespace of a CPC cluster and to the LKE cluster. It also lists secrets
which are to be copied from the kube-system namespace of the CPC to the
kube-system-<name> namespace of the CPC, and secrets to-be-copied from
the kube-system-<name> namespace of the CPC to the kube-system namespace
of the LKE. Finally, for some secrets it also defines format: the number
and names of secret items.

2. Add per-chart configuration to charts/<version>/<chart>/config.json.
This file lists resources and secrets which are required to deply this
resources.

3. The apiserver service is now deployed directly without using a chart,
as this is concerned a version-based code.

4. To review changes, it's simpler to start from the Reconcile function
and follow the logic.